Add todo list shared state container. Fix infinite items bug.
